Moved Atmos Control to separate AtmosManager
