--- UnityProject/Assets/Scripts/Editor/Matrix/AtmosSystemEditor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 0e21bc5db0774e8681d615d2a81c7ba3
-timeCreated: 1538951609
