+++ UnityProject/Assets/Scripts/Tilemaps/Behaviours/Meta/Atmospherics/AtmosManager.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+
+public class AtmosManager : MonoBehaviour
+{
+	public float Speed = 0.1f;
+
+	public int NumberThreads = 1;
+
+	private void OnValidate()
+	{
+		AtmosThread.SetSpeed(Speed);
+
+		// TODO set number of threads
+	}
+
+}
