+++ UnityProject/Assets/Scripts/Tilemaps/Behaviours/Meta/Atmospherics/AtmosSystem.cs
@@ -4,29 +4,7 @@ using UnityEngine;
 
 public class AtmosSystem : SubsystemBehaviour
 {
-	public float Speed = 0.1f;
-
 	public override void Initialize()
-	{
-		InitializeAtmos();
-	}
-
-	public override void UpdateAt(Vector3Int position)
-	{
-		AtmosThread.Enqueue(metaDataLayer.Get(position));
-	}
-
-	private void OnValidate()
-	{
-		AtmosThread.SetSpeed(Speed);
-	}
-
-	private void OnDestroy()
-	{
-		AtmosThread.Stop();
-	}
-
-	private void InitializeAtmos()
 	{
 		BoundsInt bounds = metaTileMap.GetBounds();
 
@@ -36,5 +14,10 @@ public class AtmosSystem : SubsystemBehaviour
 
 			node.Atmos = node.IsRoom ? GasMixes.Air : GasMixes.Space;
 		}
+	}
+
+	public override void UpdateAt(Vector3Int position)
+	{
+		AtmosThread.Enqueue(metaDataLayer.Get(position));
 	}
 }
